[수연][infinite & result memo] lazy-load & memo 주석
