+++ FE/src/pages/recommend/RecommendPage.js
@@ -1,5 +1,5 @@
-import { useEffect, useState } from "react";
-import { useNavigate, Link } from "react-router-dom";
+import React, { useEffect, useState, Suspense } from "react";
+import { useNavigate } from "react-router-dom";
 import { HomeFilled, SearchOutlined } from "@ant-design/icons";
 import { Select } from "antd";
 
@@ -25,10 +25,7 @@ const RecommendPage = () => {
   const mainText = "마음에 드는 여행 프로젝트를\n 내 프로젝트로! 😆";
   const [isSearchResult, setIsSearchResult] = useState(false);
 
-  // infinite scroll
-  const [uploadProjectInfo, setUploadProjectInfo] = useState([]);
-  const [skip, setSkip] = useState(0);
-
+  const RecommendRows = React.lazy(() => import("../../components/recommend/RecommendRows"));
   const children = [];
   for (let i = 0; i < hashTag.length; i++) {
     children.push(<Option key={hashTag[i] + i}>{hashTag[i]}</Option>);
@@ -62,63 +59,6 @@ const RecommendPage = () => {
     }
   }
 
-  // infinite scroll
-  useEffect(() => {
-    const fetchUploadProjectInfo = async () => {
-      try {
-        const request = await fetch(`https://${process.env.REACT_APP_SERVER_IP}:8443/recommend/infinite?skip=${skip}`);
-        const uploadProjectInfoJson = await request.json();
-        setUploadProjectInfo([...uploadProjectInfo, ...uploadProjectInfoJson]);
-        // console.log("인피니트 스크롤 결과");
-        // console.log(uploadProjectInfoJson);
-        if (uploadProjectInfoJson.length === 0) {
-          setSkip(0);
-        }
-      } catch (e) {
-        console.log("말도안돼 T_T");
-      }
-    };
-    fetchUploadProjectInfo();
-  }, [skip]);
-
-  const handleScroll = (e) => {
-    const { offsetHeight, scrollTop, scrollHeight } = e.target;
-
-    if (offsetHeight + scrollTop === scrollHeight) {
-      setSkip(uploadProjectInfo.length);
-    }
-  };
-
-  const ScrollRow = ({ el }) => {
-    const defaultHashTags = ["제주도", "여행"];
-    return (
-      <div>
-        <Link to={`project/${el._id}`}>
-          <RecommendItems>
-            <div className="background-img" style={{ backgroundImage: `url(${el.img})` }}>
-              <div className="uploadProjectInfo-title">{el.project_title}</div>
-              <div className="uploadProjectInfo-hashTags">
-                {el.hashTags.length === 0
-                  ? defaultHashTags.map((hashTag, index) => (
-                      <span key={index}>
-                        {"#"}
-                        {hashTag}{" "}
-                      </span>
-                    ))
-                  : el.hashTags.map((hashTag, index) => (
-                      <span key={index}>
-                        {"#"}
-                        {hashTag}{" "}
-                      </span>
-                    ))}
-              </div>
-            </div>
-          </RecommendItems>
-        </Link>
-      </div>
-    );
-  };
-
   return (
     <RecommendWrapper>
       <SearchBlock>
@@ -144,18 +84,10 @@ const RecommendPage = () => {
       {!isSearchResult && (
         <RecommendBlock>
           {mainText}
-          <div className="scrollWrapper">
-            <HashtagResultTextDark>🏝 모든 프로젝트</HashtagResultTextDark>
-            <RecommendContents onWheel={handleScroll}>
-              {uploadProjectInfo.map((el, i) => {
-                return i % 2 === 0 ? null : <ScrollRow key={i} el={el} />;
-              })}
-            </RecommendContents>
-            <RecommendContents onWheel={handleScroll}>
-              {uploadProjectInfo.map((el, i) => {
-                return i % 2 === 0 ? <ScrollRow key={i} el={el} /> : null;
-              })}
-            </RecommendContents>
+          <div>
+            <Suspense fallback={<div>Loading...</div>}>
+              <RecommendRows />
+            </Suspense>
           </div>
         </RecommendBlock>
       )}
@@ -204,10 +136,6 @@ const HashtagResultText = styled(HashtagResult)`
   color: #f8f9fa;
 `;
 
-const HashtagResultTextDark = styled(HashtagResultText)`
-  color: #232a3c;
-`;
-
 const RecommendWrapper = styled.div`
   background-color: #ff8a3d;
   width: 100%;
@@ -249,40 +177,6 @@ const RecommendContents = styled.div`
     /* width: 0px;
     height: 7px; */
     display: none;
-  }
-`;
-
-const RecommendItems = styled.div`
-  height: 200px;
-  width: 200px;
-  border-radius: 10px;
-  background-color: white;
-  margin-right: 20px;
-  cursor: pointer;
-  background: #232a3c;
-  
-  .background-img {
-    height: 200px;
-    width: 200px;
-    border-radius: 10px;
-    box-shadow: 4px 4px 4px rgb(0 0 0 / 25%);
-    background-position: center;
-  }
-
-  .uploadProjectInfo-title {
-    font-size: 25px;
-    color: #f8f9fa;
-    text-align: center;
-    padding-top: 40px;
-  }
-  .uploadProjectInfo-hashTags {
-    font-size: 15px;
-    color: #7c8289;
-    text-align: center;
-    margin-top: 33px;
-    background: white;
-    border-radius: 0px 0px 10px 10px;
-    height: 62px;
   }
 `;
