+++ FE/src/pages/result/Result.js
@@ -155,11 +155,11 @@ function Result() {
               </div>
             );
           })}
-        <ResultLine />
+        {/* <ResultLine />
         <ResultTitle>Memo</ResultTitle>
         <ResultMemoBox>
           <ShowMemoResult />
-        </ResultMemoBox>
+        </ResultMemoBox> */}
       </ResultContainer>
       <ResultMap routes={routes} />
     </ResultWhole>
